+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## [1.1.0]
+### Changed
+* To improve the performance of dependent picker, the multi picker will generate picker columns dynamically at run time
+### Added
+* Expose `columnWidth`, allow to configure column width manually
+
 ## [1.0.9]
 ### Added 
 * Allow to customize separator between values from different columns, ref to [issue #17](https://github.com/raychenfj/ion-multi-picker/issues/17)
@@ -8,7 +14,6 @@
 ## [1.0.8]
 ### Added
 * Migrate to ionic 2.0.0-rc3
-* Add protractor type definition
 
 ## [1.0.7]
 ### Removed